Refactor agents into packages
