+++ jarvis/agents/calendar_agent/__init__.py
@@ -2,12 +2,11 @@
 from typing import Any, Dict, Set, List
 import json
 import uuid
-from .base import NetworkAgent
-from .message import Message
-from ..services.calendar_service import CalendarService
-from ..ai_clients import BaseAIClient
-from ..logger import JarvisLogger
-from typing import Any, Dict, Set, List
+from ..base import NetworkAgent
+from ..message import Message
+from ...services.calendar_service import CalendarService
+from ...ai_clients import BaseAIClient
+from ...logger import JarvisLogger
 from datetime import datetime, timedelta, timezone
